Implemented CSV upload and board importing
